Make it possible to always cast up (safe), never cast down (uncertain).
Casting up from usize to u64 is safe, but casting down is not.

So, change types of u64 constants that are cast down to usize to usize and
cast up to u64 where necessary.

It is easy to tell by observation where the constant is defined that it fits
into the smaller type.

Constants affected:
* SECTOR_SIZE
* BDA_STATIC_HDR_SIZE
* NUM_PRIMARY_MDA_REGIONS

+++ src/consts.rs
@@ -2,7 +2,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-pub const SECTOR_SIZE: u64 = 512;
+pub const SECTOR_SIZE: usize = 512;
 
 pub const MEGA: u64 = 1024 * 1024;
 pub const GIGA: u64 = MEGA * 1024;
